requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
openai==0.28.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
nltk==3.8.1
pydub==0.25.1
gTTS==2.3.2